Improved coding style
