+++ src/actions/artefact.js
@@ -3,9 +3,6 @@ import {
     REGISTER_ARTEFACT_REQUEST,
     REGISTER_ARTEFACT_SUCCESS,
     REGISTER_ARTEFACT_FAILURE,
-    REGISTER_GROUP_REQUEST,
-    REGISTER_GROUP_SUCCESS,
-    REGISTER_GROUP_FAILURE,
   } from 'actions/types';
   
   /**
@@ -38,25 +35,3 @@ import {
   });
   
 
-    /**
-   * @route CreateGroup
-   * @method POST
-   * @param {string} groupName
-   * @param {ImageBitmap} coverPhoto
-   * @param {string} details
-   * @param {String} tag
-   * @desc create an group for  user
-   */
-  export const registerGroup = (groupName, coverPhoto, details, tag) => ({
-    [CALL_API]: {
-      endpoint: '/CreateGroup',
-      method: 'POST',
-      body: {
-        groupName: groupName,
-        coverPhoto: coverPhoto,
-        details: details,
-        tag: tag,
-      },
-      types: [REGISTER_GROUP_REQUEST, REGISTER_GROUP_SUCCESS, REGISTER_GROUP_FAILURE],
-    },
-  });
